+++ open_with/config.py
@@ -17,13 +17,13 @@ try:
                         QLineEdit, QTableWidget, QTableWidgetItem, QFileDialog,
                         QAbstractItemView, QRadioButton, QAction, QIcon, QToolButton,
                         QDialog, QDialogButtonBox, QGridLayout, QGroupBox,
-                        QInputDialog, QSpacerItem)
+                        QInputDialog, QSpacerItem, QModelIndex)
 except ImportError:
     from PyQt5.Qt import (Qt, QWidget, QVBoxLayout, QHBoxLayout, QLabel, QPushButton,
                         QLineEdit, QTableWidget, QTableWidgetItem, QFileDialog,
                         QAbstractItemView, QRadioButton, QAction, QIcon, QToolButton,
                         QDialog, QDialogButtonBox, QGridLayout, QGroupBox,
-                        QInputDialog, QSpacerItem)
+                        QInputDialog, QSpacerItem, QModelIndex)
 
 try:
     load_translations()
@@ -390,20 +390,23 @@ class MenuTableWidget(QTableWidget):
 
     def delete_rows(self):
         self.setFocus()
-        rows = self.selectionModel().selectedRows()
-        if len(rows) == 0:
+        selrows = self.selectionModel().selectedRows()
+        selrows = sorted(selrows, key=lambda x: x.row())
+        if len(selrows) == 0:
             return
-        message = '<p>'+_('Are you sure you want to delete this menu item?')
-        if len(rows) > 1:
-            message = '<p>'+_('Are you sure you want to delete the selected %d menu items?')%len(rows)
-        if not question_dialog(self, _('Are you sure?'), message, show_copy_button=False):
+        message = _('Are you sure you want to delete this menu item?')
+        if len(selrows) > 1:
+            message = _('Are you sure you want to delete the selected {0} menu items?').format(len(selrows))
+        if not question_dialog(self, _('Are you sure?'), '<p>'+message, show_copy_button=False):
             return
-        first_sel_row = self.currentRow()
-        for selrow in reversed(rows):
-            self.removeRow(selrow.row())
-        if first_sel_row < self.rowCount():
+        first_sel_row = selrows[0].row()
+        for selrow in reversed(selrows):
+            self.model().removeRow(selrow.row())
+        if first_sel_row < self.model().rowCount(QModelIndex()):
+            self.setCurrentIndex(self.model().index(first_sel_row, 0))
             self.select_and_scroll_to_row(first_sel_row)
-        elif self.rowCount() > 0:
+        elif self.model().rowCount(QModelIndex()) > 0:
+            self.setCurrentIndex(self.model().index(first_sel_row - 1, 0))
             self.select_and_scroll_to_row(first_sel_row - 1)
 
     def move_rows_up(self):