+++ resize_cover/config.py
@@ -10,11 +10,11 @@ from six import text_type as unicode
 try:
     from qt.core import (Qt, QWidget, QVBoxLayout, QHBoxLayout, QLabel, QTableWidget,
                       QPushButton, QTableWidgetItem, QIcon, QAbstractItemView,
-                      QToolButton, QSpacerItem)
+                      QToolButton, QSpacerItem, QModelIndex)
 except ImportError:
     from PyQt5.Qt import (Qt, QWidget, QVBoxLayout, QHBoxLayout, QLabel, QTableWidget,
                       QPushButton, QTableWidgetItem, QIcon, QAbstractItemView,
-                      QToolButton, QSpacerItem)
+                      QToolButton, QSpacerItem, QModelIndex)
 
 from calibre.gui2 import question_dialog, info_dialog
 from calibre.utils.config import JSONConfig
@@ -137,23 +137,26 @@ class SizesTableWidget(QTableWidget):
 
     def delete_rows(self):
         self.setFocus()
-        rows = self.selectionModel().selectedRows()
-        if len(rows) == 0:
+        selrows = self.selectionModel().selectedRows()
+        selrows = sorted(selrows, key=lambda x: x.row())
+        if len(selrows) == 0:
             return
-        if self.rowCount() == 1 or self.rowCount() == len(rows):
+        if self.rowCount() == 1 or self.rowCount() == len(selrows):
             return info_dialog(self, _('Cannot delete'), _('You must have at least one resize menu item'))
 
         message = '<p>' + _('Are you sure you want to delete this resize menu item?')
-        if len(rows) > 1:
-            message = '<p>' + _('Are you sure you want to delete the selected {0} resize menu items?').format(len(rows))
+        if len(selrows) > 1:
+            message = '<p>' + _('Are you sure you want to delete the selected {0} resize menu items?').format(len(selrows))
         if not question_dialog(self, _('Are you sure?'), message, show_copy_button=False):
             return
-        first_sel_row = self.currentRow()
-        for selrow in reversed(rows):
-            self.removeRow(selrow.row())
-        if first_sel_row < self.rowCount():
+        first_sel_row = selrows[0].row()
+        for selrow in reversed(selrows):
+            self.model().removeRow(selrow.row())
+        if first_sel_row < self.model().rowCount(QModelIndex()):
+            self.setCurrentIndex(self.model().index(first_sel_row, 0))
             self.select_and_scroll_to_row(first_sel_row)
-        elif self.rowCount() > 0:
+        elif self.model().rowCount(QModelIndex()) > 0:
+            self.setCurrentIndex(self.model().index(first_sel_row - 1, 0))
             self.select_and_scroll_to_row(first_sel_row - 1)
 
     def set_as_default(self):