+++ search_the_internet/config.py
@@ -20,13 +20,13 @@ try:
                           QIcon, QFormLayout, QAction, QFileDialog, QDialog, QTableWidget,
                           QTableWidgetItem, QAbstractItemView, QComboBox,
                           QGroupBox, QGridLayout, QRadioButton, QDialogButtonBox,
-                          QPushButton, QToolButton, QSpacerItem)
+                          QPushButton, QToolButton, QSpacerItem, QModelIndex)
 except:
     from PyQt5.Qt import (Qt, QWidget, QVBoxLayout, QHBoxLayout, QLabel, QLineEdit,
                           QIcon, QFormLayout, QAction, QFileDialog, QDialog, QTableWidget,
                           QTableWidgetItem, QAbstractItemView, QComboBox,
                           QGroupBox, QGridLayout, QRadioButton, QDialogButtonBox,
-                          QPushButton, QToolButton, QSpacerItem)
+                          QPushButton, QToolButton, QSpacerItem, QModelIndex)
 
 from calibre.ebooks.metadata.book.base import Metadata
 from calibre.gui2 import (error_dialog, question_dialog, info_dialog, choose_files,
@@ -701,20 +701,23 @@ class MenuTableWidget(QTableWidget):
 
     def delete_rows(self):
         self.setFocus()
-        rows = self.selectionModel().selectedRows()
-        if len(rows) == 0:
+        selrows = self.selectionModel().selectedRows()
+        selrows = sorted(selrows, key=lambda x: x.row())
+        if len(selrows) == 0:
             return
-        message = _('<p>Are you sure you want to delete this menu item?')
-        if len(rows) > 1:
-            message = _('<p>Are you sure you want to delete the selected {0} menu items?').format(len(rows))
-        if not question_dialog(self, _('Are you sure?'), message, show_copy_button=False):
+        message = _('Are you sure you want to delete this menu item?')
+        if len(selrows) > 1:
+            message = _('Are you sure you want to delete the selected {0} menu items?').format(len(selrows))
+        if not question_dialog(self, _('Are you sure?'), '<p>'+message, show_copy_button=False):
             return
-        first_sel_row = self.currentRow()
-        for selrow in reversed(rows):
-            self.removeRow(selrow.row())
-        if first_sel_row < self.rowCount():
+        first_sel_row = selrows[0].row()
+        for selrow in reversed(selrows):
+            self.model().removeRow(selrow.row())
+        if first_sel_row < self.model().rowCount(QModelIndex()):
+            self.setCurrentIndex(self.model().index(first_sel_row, 0))
             self.select_and_scroll_to_row(first_sel_row)
-        elif self.rowCount() > 0:
+        elif self.model().rowCount(QModelIndex()) > 0:
+            self.setCurrentIndex(self.model().index(first_sel_row - 1, 0))
             self.select_and_scroll_to_row(first_sel_row - 1)
 
     def move_rows_up(self):